Generate wasi worker
