+++ cli/package.json
@@ -80,6 +80,7 @@
     "typanion": "^3.13.0"
   },
   "devDependencies": {
+    "@emnapi/core": "0.43.1",
     "@emnapi/runtime": "0.43.1",
     "@types/debug": "^4.1.7",
     "@types/inquirer": "^9.0.3",
