--- wasi-runtime/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "composite": true,
-    "outDir": "dist",
-    "rootDir": "./src",
-    "types": ["node"]
-  },
-  "include": ["src"]
-}
